+++ src/Language/AST/InterfaceTypeExtensionNode.php
@@ -15,7 +15,7 @@ class InterfaceTypeExtensionNode extends Node implements TypeExtensionNode
     /** @var NodeList<DirectiveNode> */
     public $directives;
 
-    /** @var NodeList<InterfaceTypeDefinitionNode> */
+    /** @var NodeList<NamedTypeNode> */
     public $interfaces;
 
     /** @var NodeList<FieldDefinitionNode> */